+++ packages/basehub/next-toolbar.d.ts
@@ -1,2 +1,1 @@
-/* eslint-disable import/no-unresolved */
-export * from "./dist/generated-client/next-toolbar.d.ts";
+export * from "../../playground/.basehub/next-toolbar/index.d.ts";