+++ packages/basehub/react-pump.d.ts
@@ -1,2 +1,1 @@
-/* eslint-disable import/no-unresolved */
-export * from "./dist/generated-client/react-pump";
+export * from "../../playground/.basehub/react-pump/index.d.ts";