+++ packages/basehub/src/react/index.ts
@@ -3,5 +3,5 @@ export {
   createRichTextWithDefaultComponents,
   type RichTextProps,
   type CustomBlocksBase,
-  type HandlerProps,
+  type HandlerProps as RichTextHandlerProps,
 } from "./rich-text/primitive";