+++ packages/basehub/src/react/pump/server-pump.tsx
@@ -41,7 +41,6 @@ export type PumpProps<Queries extends Array<PumpQuery>> = {
     data: QueryResults<Queries>
   ) => React.ReactNode | Promise<React.ReactNode>;
   queries: [...Queries]; // Tuple type for better type inference
-  disableNextjsAutoDraftMode?: boolean;
 } & Parameters<typeof basehub>[0];
 
 // Utility type to infer result types from an array of queries
@@ -52,7 +51,6 @@ export type QueryResults<Queries extends Array<PumpQuery>> = {
 export const Pump = async <Queries extends Array<PumpQuery>>({
   children,
   queries,
-  disableNextjsAutoDraftMode,
   ...basehubProps
 }: PumpProps<Queries>) => {
   // passed to the client to toast
@@ -60,7 +58,8 @@ export const Pump = async <Queries extends Array<PumpQuery>>({
   const responseHashes: Array<ResponseCache["responseHash"]> = [];
 
   let isNextjsDraftMode = false;
-  if (!disableNextjsAutoDraftMode) {
+  if (basehubProps.draft === undefined) {
+    // try to auto-detect (only if draft is not explicitly set by the user)
     try {
       const { draftMode } = await import("next/headers");
       isNextjsDraftMode = (await draftMode()).isEnabled;
@@ -82,6 +81,20 @@ export const Pump = async <Queries extends Array<PumpQuery>>({
 
   const queriesWithFallback =
     draft && noQueries ? [{ _sys: { id: true } }] : queries;
+
+  if (draft) {
+    // try to get ref from cookies
+    try {
+      const { cookies } = await import("next/headers");
+      const cookieStore = await cookies();
+      const ref = cookieStore.get("bshb-preview-ref")?.value;
+      if (ref) {
+        headers["x-basehub-ref"] = ref;
+      }
+    } catch (error) {
+      // noop
+    }
+  }
 
   const results: Array<{
     data: QueryResults<Queries>[number] | undefined;
@@ -199,7 +212,7 @@ export const Pump = async <Queries extends Array<PumpQuery>>({
         pumpToken={pumpToken ?? undefined}
         initialResolvedChildren={resolvedChildren}
         apiVersion={apiVersion}
-        resolvedRef={resolvedRef}
+        previewRef={headers["x-basehub-ref"] || resolvedRef.ref}
       >
         {/* We pass the raw `children` param as it might be a server action that will be re-executed from the client as data comes in */}
         {/* @ts-ignore */}