+++ packages/basehub/tsup-client.config.ts
@@ -8,11 +8,11 @@ export default defineConfig((_options: Options) => {
     dts: true,
     entry: {
       "react-rich-text": "./src/react/rich-text/index.ts",
+      "react-form": "./src/react/form/index.ts",
       "react-code-block/index": "./src/react/code-block/index.ts",
       "react-code-block/client": "./src/react/code-block/client/index.ts",
       "react-search": "./src/react/search/index.ts",
       "api-transaction": "./src/api-transaction.ts",
-      analytics: "./src/analytics/index.ts",
       search: "./src/search/index.ts",
       "next-image": "./src/next/image/index.ts",
     },