+++ playground/CHANGELOG.md
@@ -1,5 +1,7 @@
 # playground
 
+## 0.0.164-canary.17
+
 ## 0.0.170
 
 ### Patch Changes
@@ -40,7 +42,105 @@
 ### Patch Changes
 
 - Updated dependencies
-  - basehub@7.5.23
+  - basehub@8.0.0-canary.17
+
+## 0.0.164-canary.16
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.16
+
+## 0.0.164-canary.15
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.15
+
+## 0.0.164-canary.14
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.14
+
+## 0.0.164-canary.13
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.13
+
+## 0.0.164-canary.12
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.12
+
+## 0.0.164-canary.11
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.11
+
+## 0.0.164-canary.10
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.10
+
+## 0.0.164-canary.9
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.9
+
+## 0.0.164-canary.8
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.8
+
+## 0.0.164-canary.7
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.7
+
+## 0.0.164-canary.6
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.6
+
+## 0.0.164-canary.5
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.5
+
+## 0.0.164-canary.4
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.4
+
+## 0.0.164-canary.3
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.3
 
 ## 0.0.164
 
@@ -64,21 +164,21 @@
 ### Patch Changes
 
 - Updated dependencies
-  - basehub@7.5.22-canary.2
+  - basehub@8.0.0-canary.2
 
 ## 0.0.164-canary.1
 
 ### Patch Changes
 
 - Updated dependencies
-  - basehub@7.5.22-canary.1
+  - basehub@8.0.0-canary.1
 
 ## 0.0.164-canary.0
 
 ### Patch Changes
 
 - Updated dependencies
-  - basehub@7.5.22-canary.0
+  - basehub@8.0.0-canary.0
 
 ## 0.0.163
 
@@ -1202,8 +1302,6 @@
 
 - Updated dependencies
   - basehub@4.0.16-canary.0
-
-<<<<<<< HEAD
 
 ## 0.0.83-next.0